feat(auth): rename register to signup and update related routes
