+++ src/app/login/page.tsx
@@ -89,7 +89,7 @@ export default function Login() {
         </div>
         <p className="text-sm text-center text-gray-600">
           Don&apos;t have an account?{' '}
-          <Link href="/register" className="text-blue-600 hover:underline">
+          <Link href="/sign-up" className="text-blue-600 hover:underline">
             Sign up
           </Link>
         </p>
