+++ src/app/sign-up/page.tsx
@@ -1,0 +1,132 @@
+'use client'
+import Link from "next/link";
+import { Input } from "@/components/ui/input";
+import { Button } from "@/components/ui/button";
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import { signup } from "@/api/auth";
+import { signIn } from "next-auth/react";
+
+
+export default function SignUp() {
+  const router = useRouter();
+  const [step, setStep] = useState(1); // 追蹤當前步驟
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleNextStep = (e: React.FormEvent) => {
+    e.preventDefault();
+    setError("");
+    if (!name || !email) {
+      setError("Please fill in all fields");
+      return;
+    }
+    setStep(2);
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setError("");
+    if (!password) {
+      setError("Please fill in all fields");
+      return;
+    }
+    setIsLoading(true);
+
+    try {
+      await signup(name, email, password);
+
+      // 自動登入
+      const signInRes = await signIn("credentials", {
+        redirect: false,
+        email,
+        password,
+      });
+
+      if (signInRes?.error) {
+        setError(signInRes.error);
+      } else {
+        router.push("/");
+      }
+    } catch (err: Error | unknown) {
+      const errorMessage = err instanceof Error ? err.message : "sign up failed";
+      setError(errorMessage);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <div className="flex items-center justify-center bg-gray-50 min-h-[calc(100vh-4rem-1px)]">
+      <div className="w-full max-w-md p-8 space-y-4 bg-white rounded-lg shadow-md">
+        <h2 className="text-2xl font-bold text-center text-gray-900">Sign up</h2>
+        <p className="text-center text-gray-600">
+          {step === 1 ? "Enter your name and email to continue" : "Enter your password to complete registration"}
+        </p>
+        {error && <p className="text-sm text-red-600">{error}</p>}
+        {step === 1 && (
+          <form className="space-y-4" onSubmit={handleNextStep}>
+            <div>
+              <label htmlFor="name" className="block text-sm font-medium text-gray-700">
+                Name
+              </label>
+              <Input
+                id="name"
+                type="text"
+                placeholder="John Doe"
+                className="mt-1"
+                value={name}
+                onChange={e => setName(e.target.value)}
+              />
+            </div>
+            <div>
+              <label htmlFor="email" className="block text-sm font-medium text-gray-700">
+                Email
+              </label>
+              <Input
+                id="email"
+                type="email"
+                placeholder="m@example.com"
+                className="mt-1"
+                value={email}
+                onChange={e => setEmail(e.target.value)}
+              />
+            </div>
+            <Button type="submit" className="w-full" disabled={isLoading}>
+              {isLoading ? "Loading..." : "Continue with email"}
+            </Button>
+          </form>
+        )}
+        {step === 2 && (
+          <form className="space-y-4" onSubmit={handleSubmit}>
+            <div>
+              <label htmlFor="password" className="block text-sm font-medium text-gray-700">
+                Password
+              </label>
+              <Input
+                id="password"
+                type="password"
+                placeholder="Enter your password"
+                className="mt-1"
+                value={password}
+                onChange={e => setPassword(e.target.value)}
+              />
+            </div>
+            <Button type="submit" className="w-full" disabled={isLoading}>
+              {isLoading ? "Loading..." : "Sign up"}
+            </Button>
+          </form>
+        )}
+        <p className="text-sm text-center text-gray-600">
+          Already have an account?{' '}
+          <Link href="/login" className="text-blue-600 hover:underline">
+            Log in
+          </Link>
+        </p>
+      </div>
+    </div>
+  );
+}
