+++ src/components/site-header.tsx
@@ -65,8 +65,8 @@ export function SiteHeader() {
                 <Link href="/login" className="inline-flex items-center justify-center rounded-md px-4 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground h-10">
                   Login
                 </Link>
-                <Link href="/register" className="inline-flex items-center justify-center rounded-md px-4 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground h-10">
-                  Register
+                <Link href="/sign-up" className="inline-flex items-center justify-center rounded-md px-4 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground h-10">
+                  Sign up
                 </Link>
               </>
             )}
